+++ .github/workflows/ci.yml
@@ -55,30 +55,20 @@ jobs:
 
 # 3rd team #########################################################
 
-    - name: "[blst-from-scratch] Tests"
-      uses: actions-rs/cargo@v1
+    - uses: actions-rs/cargo@v1
       with:
         command: test
-        args: --manifest-path blst-from-scratch/Cargo.toml
+        args: --manifest-path from-scratch/Cargo.toml
 
-    - name: "[blst-from-scratch] Benchmark"
-      uses: actions-rs/cargo@v1
+    - uses: actions-rs/cargo@v1
       with:
         command: bench
-        args: --manifest-path blst-from-scratch/Cargo.toml
+        args: --manifest-path from-scratch/Cargo.toml
 
-#  - name: "[blst-from-scratch] Clippy"
-#     uses: actions-rs/cargo@v1
-#     with:
-#     command: clippy
-#     args: --manifest-path blst-from-scratch/Cargo.toml --all-targets --all-features -- -D warnings
-
-# Incompatible with Windows CI
-#    - name: "[blst-from-scratch] Formatting"
-#      uses: actions-rs/cargo@v1
-#      with:
-#        command: fmt
-#        args: --manifest-path blst-from-scratch/Cargo.toml
+    - uses: actions-rs/cargo@v1
+      with:
+        command: fmt
+        args: --manifest-path from-scratch/Cargo.toml
 
 # 4th team #########################################################
 
@@ -126,7 +116,7 @@ jobs:
       with:
         command: clippy
         args: --manifest-path mcl-kzg/kzg/Cargo.toml --all-targets --all-features -- -D warnings
-
+        
     - name: "Run mcl-kzg10-rust bench Clippy"
       uses: actions-rs/cargo@v1
       with:
@@ -136,18 +126,18 @@ jobs:
     - name: Run mcl-kzg10-rust tests
       if: matrix.os == 'ubuntu-latest'
       run: |
-        cargo test shared_tests --manifest-path mcl-kzg/kzg-bench/Cargo.toml -- --test-threads 1
-
+        cargo test shared_tests --manifest-path mcl-kzg/kzg-bench/Cargo.toml -- --test-threads 1 
+        
     - name: Run mcl-kzg10-rust benches
       if: matrix.os == 'ubuntu-latest'
       run: |
         cargo bench --manifest-path mcl-kzg/kzg-bench/Cargo.toml
-
+        
     - name: Run mcl-kzg10-rust fmt
       if: matrix.os == 'ubuntu-latest'
       run: |
         cargo fmt --manifest-path mcl-kzg/kzg-bench/Cargo.toml
-
+        
 # Common ###########################################################
 
     - uses: actions-rs/cargo@v1
