--- blst-from-scratch/README.md
@@ -1,46 +1,0 @@
-# KZG10-BLST
-## [BLS12-381 lib](https://github.com/supranational/blst)
-
-Planned functionality:
- * Finite field FFT (Fast Fourier Transform);
- * G1 group FFT;
- * Data availability sampling (data extension, recovery from any 1/2 samples);
- * Zero polynomials;
- * KZG10 single commit/verify;
- * KZG10 multi commit/verify;
- * FK20 single proof regular/optimised for DAS; 
- * FK20 multi proof regular/optimised for DAS.
- * Test suite (based on C implementation)
- * Benchmarks
-
-## Other implementations
-* [C implementation](https://github.com/benjaminion/c-kzg)
-* [Go implementation](https://github.com/protolambda/go-kzg)
-* [Rust herumi implementation](https://github.com/UndeadRat22/kzg10-rust)
-
-```
-     .-.            .-.
-    /   \          /   \
-   |   _ \        / _   |
-   ;  | \ \      / / |  ;
-    \  \ \ \_.._/ / /  /
-     '. '.;'    ';,' .'
-       './ _    _ \.'
-       .'  a __ a  '.
-  '--./ _,   \/   ,_ \.--'
- ----|   \   /\   /   |----
-  .--'\   '-'  '-'    /'--.
-      _>.__  -- _.-  `;
-    .' _     __/     _/
-   /    '.,:".-\    /:,
-   |      \.'   `""`'.\\
-    '-,.__/  _   .-.  ;|_
-    /` `|| _/ `\/_  \_|| `\
-   |    ||/ \-./` \ / ||   |
-    \   ||__/__|___|__||  /
-     \_ |__BLS12-381__| /
-    .'  \ =  _= _ = _= /`\
-   /     `-;----=--;--'   \
-   \    _.-'        '.    /
-    `""` Happy Easter `""`
-```
