--- blst-from-scratch/tests/fk20_proofs.rs
@@ -1,36 +1,0 @@
-#[cfg(test)]
-mod tests {
-    use kzg_bench::tests::fk20_proofs::*;
-    use blst_from_scratch::kzg_types::{FsFFTSettings, FsFK20MultiSettings, FsFK20SingleSettings, FsFr, FsG1, FsG2, FsKZGSettings, FsPoly};
-    use blst_from_scratch::utils::generate_trusted_setup;
-
-    #[test]
-    fn test_fk_single() {
-        fk_single::<FsFr, FsG1, FsG2, FsPoly, FsFFTSettings, FsKZGSettings, FsFK20SingleSettings>(&generate_trusted_setup);
-    }
-
-    #[test]
-    fn test_fk_single_strided() {
-        fk_single_strided::<FsFr, FsG1, FsG2, FsPoly, FsFFTSettings, FsKZGSettings, FsFK20SingleSettings>(&generate_trusted_setup);
-    }
-
-    #[test]
-    fn test_fk_multi_settings() {
-        fk_multi_settings::<FsFr, FsG1, FsG2, FsPoly, FsFFTSettings, FsKZGSettings, FsFK20MultiSettings>(&generate_trusted_setup);
-    }
-
-    #[test]
-    fn test_fk_multi_chunk_len_1_512() {
-        fk_multi_chunk_len_1_512::<FsFr, FsG1, FsG2, FsPoly, FsFFTSettings, FsKZGSettings, FsFK20MultiSettings>(&generate_trusted_setup);
-    }
-
-    #[test]
-    fn test_fk_multi_chunk_len_16_512() {
-        fk_multi_chunk_len_16_512::<FsFr, FsG1, FsG2, FsPoly, FsFFTSettings, FsKZGSettings, FsFK20MultiSettings>(&generate_trusted_setup);
-    }
-
-    #[test]
-    fn test_fk_multi_chunk_len_16_16() {
-        fk_multi_chunk_len_16_16::<FsFr, FsG1, FsG2, FsPoly, FsFFTSettings, FsKZGSettings, FsFK20MultiSettings>(&generate_trusted_setup);
-    }
-}
