--- blst-from-scratch/tests/local_tests/mod.rs
@@ -1,3 +1,0 @@
-pub mod local_poly;
-pub mod local_consts;
-pub mod local_recovery;
