--- blst-from-scratch/tests/mod.rs
@@ -1,1 +1,0 @@
-pub mod local_tests;
