+++ from-scratch/README.md
@@ -1,0 +1,47 @@
+# KZG10-BLST
+## [BLS12-381 lib](https://github.com/supranational/blst)
+
+Planned functionality:
+ * Finite field FFT (Fast Fourier Transform);
+ * G1 group FFT;
+ * Data availability sampling (data extension, recovery from any 1/2 samples);
+ * Zero polynomials;
+ * KZG10 single commit/verify;
+ * KZG10 multi commit/verify;
+ * FK20 single proof regular/optimised for DAS; 
+ * FK20 multi proof regular/optimised for DAS.
+ * Test suite (based on C implementation)
+ * Benchmarks
+
+## Other implementations
+* [C implementation](https://github.com/benjaminion/c-kzg)
+* [Go implementation](https://github.com/protolambda/go-kzg)
+* [Rust herumi implementation](https://github.com/UndeadRat22/kzg10-rust)
+
+```
+           *             ,
+                       _/^\_
+                      < OwO >
+     *                 /.-.\         *
+              *        `/&\`                   *
+                      ,@.*;@,
+                     /_o.I %_\    *
+        *           (`'--:o(_@;
+                   /`;--.,__ `')             *
+                  ;@`o % O,*`'`&\
+            *    (`'--)_@ ;o %'()\      *
+                 /`;--._`''--._O'@;
+                /&*,()~o`;-.,_ `""`)
+     *          /`,@ ;+& () o*`;-';\
+               (`""--.,_0 +% @' &()\
+               /-.,_    ``''--....-'`)  *
+          *    /@%;o`:;'--,.__   __.'\
+              ;*,&(); @ % &^;~`"`o;@();         *
+              /(); o^~; & ().o@*&`;&%O\
+              `"="==""==,,,.,="=="==="`
+           __.----.---''#####--.___\\db///___.----._
+         '`            `"""""` |\ BLS12-381\\|
+                               |\\  \\  \\  \|
+                               |_\\__\\__\\__|
+         
+```
