+++ from-scratch/src/lib.rs
@@ -5,7 +5,7 @@ pub mod fft_g1;
 pub mod fk20_proofs;
 pub mod kzg_proofs;
 pub mod kzg_types;
+pub mod poly_utils;
 pub mod recovery;
 pub mod utils;
 pub mod zero_poly;
-pub mod bytes;
