+++ from-scratch/tests/consts.rs
@@ -1,20 +1,11 @@
-#[path = "./local_tests/local_consts.rs"]
-pub mod local_consts;
-
 #[cfg(test)]
 pub mod tests {
     use kzg_bench::tests::consts::{
         expand_roots_is_plausible, new_fft_settings_is_plausible, roots_of_unity_are_plausible,
-        roots_of_unity_out_of_bounds_fails, roots_of_unity_is_the_expected_size,
+        roots_of_unity_out_of_bounds_fails,
     };
-    use blst_from_scratch::consts::{expand_root_of_unity, SCALE2_ROOT_OF_UNITY};
-    use blst_from_scratch::kzg_types::{FsFFTSettings, FsFr};
-    use crate::local_consts::roots_of_unity_repeat_at_stride;
-
-    #[test]
-    fn roots_of_unity_is_the_expected_size_() {
-        roots_of_unity_is_the_expected_size(&SCALE2_ROOT_OF_UNITY);
-    }
+    use kzg_from_scratch::consts::{expand_root_of_unity, SCALE2_ROOT_OF_UNITY};
+    use kzg_from_scratch::kzg_types::{FsFFTSettings, FsFr};
 
     #[test]
     fn roots_of_unity_out_of_bounds_fails_() {
@@ -34,10 +25,5 @@ pub mod tests {
     #[test]
     fn new_fft_settings_is_plausible_() {
         new_fft_settings_is_plausible::<FsFr, FsFFTSettings>();
-    }
-
-    #[test]
-    fn roots_of_unity_repeat_at_stride_() {
-        roots_of_unity_repeat_at_stride::<FsFr, FsFFTSettings>();
     }
 }
