+++ from-scratch/tests/fft_fr.rs
@@ -1,8 +1,8 @@
 #[cfg(test)]
 pub mod tests {
     use kzg_bench::tests::fft_fr::{compare_sft_fft, inverse_fft, roundtrip_fft, stride_fft};
-    use blst_from_scratch::fft_fr::{fft_fr_fast, fft_fr_slow};
-    use blst_from_scratch::kzg_types::{FsFFTSettings, FsFr};
+    use kzg_from_scratch::fft_fr::{fft_fr_fast, fft_fr_slow};
+    use kzg_from_scratch::kzg_types::{FsFFTSettings, FsFr};
 
     #[test]
     fn compare_sft_fft_() {
