+++ from-scratch/tests/zero_poly.rs
@@ -4,7 +4,7 @@ pub mod tests {
         check_test_data, reduce_partials_random, test_reduce_partials, zero_poly_252,
         zero_poly_all_but_one, zero_poly_known, zero_poly_random,
     };
-    use blst_from_scratch::kzg_types::{FsFFTSettings, FsFr, FsPoly};
+    use kzg_from_scratch::kzg_types::{FsFFTSettings, FsFr, FsPoly};
 
     #[test]
     fn test_reduce_partials_() {
