+++ kzg-bench/src/benches/mod.rs
@@ -4,4 +4,3 @@ pub mod das;
 pub mod zero_poly;
 pub mod kzg;
 pub mod recover;
-pub mod fk20;
