+++ kzg-bench/src/tests/bls12_381.rs
@@ -14,6 +14,7 @@ pub fn fr_is_null_works<TFr: Fr>() {
     let null = TFr::null();
     assert!(!zero.is_null());
     assert!(!one.is_null());
+    //println!("tttttttt {}", null.l[1]);
     assert!(null.is_null());
 }
